+++ gatsby-ssr.js
@@ -8,14 +8,14 @@ import Terser from 'terser';
 
 import App from './src/components/App';
 
-import { COLOR_MODE_KEY, INITIAL_COLOR_MODE_CSS_PROP } from './src/constants';
-import { colors } from './src/theme';
+import { COLOR_MODE_KEY, INITIAL_COLOR_MODE_CSS_PROP, ENABLE_LOCAL_STORAGE } from './src/constants';
+import { COLORS } from './src/theme';
 
 // Thanks to Joshua Comeau for the original code, licensed under MIT License:
 // https://github.com/joshwcomeau/dark-mode-minimal
 
-const setColorsByTheme = () => {
-  const colorsKey = '🌈';
+function setColorsByTheme() {
+  const colors = '🌈';
   const colorModeKey = '🔑';
   const colorModeCssProp = '⚡️';
 
@@ -33,21 +33,20 @@ const setColorsByTheme = () => {
     colorMode = prefersDarkFromMQ ? 'dark' : 'light';
   }
 
-  // TODO Change to 'let' if seems like the 'const' breaks something
   const root = document.documentElement;
 
   root.style.setProperty(colorModeCssProp, colorMode);
 
-  Object.entries(colorsKey).forEach(([name, colorByTheme]) => {
+  Object.entries(colors).forEach(([name, colorByTheme]) => {
     const cssVarName = `--color-${name}`;
 
     root.style.setProperty(cssVarName, colorByTheme[colorMode]);
   });
-};
+}
 
 const MagicScriptTag = () => {
   const boundFn = String(setColorsByTheme)
-    .replace("'🌈'", JSON.stringify(colors))
+    .replace("'🌈'", JSON.stringify(COLORS))
     .replace('🔑', COLOR_MODE_KEY)
     .replace('⚡️', INITIAL_COLOR_MODE_CSS_PROP);
 
@@ -55,26 +54,39 @@ const MagicScriptTag = () => {
 
   calledFunction = Terser.minify(calledFunction).code;
 
-  // eslint-disable-next-line react/no-danger, react/jsx-filename-extension
+  // eslint-disable-next-line react/no-danger
   return <script dangerouslySetInnerHTML={{ __html: calledFunction }} />;
 };
 
+/**
+ * If the user has JS disabled, the injected script will never fire!
+ * This means that they won't have any colors set, everything will be default
+ * black and white.
+ * We can solve for this by injecting a `<style>` tag into the head of the
+ * document, which sets default values for all of our colors.
+ * Only light mode will be available for users with JS disabled.
+ */
 const FallbackStyles = () => {
-  const cssVariableString = Object.entries(colors).reduce(
-    (acc, [name, colorByTheme]) => `${acc}\n--color-${name}: ${colorByTheme.light};`,
-    '',
-  );
+  // Create a string holding each CSS variable:
+  /*
+    `--color-text: black;
+    --color-background: white;`
+  */
+
+  const cssVariableString = Object.entries(COLORS).reduce((acc, [name, colorByTheme]) => {
+    return `${acc}\n--color-${name}: ${colorByTheme.light};`;
+  }, '');
 
   const wrappedInSelector = `html { ${cssVariableString} }`;
 
   return <style>{wrappedInSelector}</style>;
 };
 
-const onRenderBody = ({ setPreBodyComponents, setHeadComponents }) => {
+export const onRenderBody = ({ setPreBodyComponents, setHeadComponents }) => {
   setHeadComponents(<FallbackStyles />);
   setPreBodyComponents(<MagicScriptTag />);
 };
 
-const wrapRootElement = ({ element }) => <App>{element}</App>;
-
-export default { onRenderBody, wrapRootElement };
+export const wrapPageElement = ({ element }) => {
+  return <App>{element}</App>;
+};