+++ src/components/Theme.jsx
@@ -6,6 +6,4 @@ import { ThemeProvider } from 'styled-components';
 
 import theme from '../theme';
 
-export default (props) => {
-  return <ThemeProvider theme={theme}>{props.children}</ThemeProvider>;
-};
+export default (props) => <ThemeProvider theme={theme}>{props.children}</ThemeProvider>;