+++ src/components/ThemeContext.jsx
@@ -4,8 +4,6 @@
 // Thanks to Joshua Comeau for the original code, licensed under MIT License:
 // https://github.com/joshwcomeau/dark-mode-minimal
 
-import React from 'react';
+import { createContext } from 'react';
 
-import ThemeContextProvider from './ThemeContextProvider';
-
-export default (props) => <ThemeContextProvider>{props.children}</ThemeContextProvider>;
+export default createContext();